Update upgrading.md
In the version requirements section, the statement `- To upgrade your cluster to v5.4, your cluster must first be on 5.0.2-30 or above.` is not accurate.  Upgrading from 5.0.2-20 to 5.4.14-28 failed with "ImportError: No module named typing" error.  This was reported in JIRA https://redislabs.atlassian.net/browse/RED-39622.  The customer must be on 5.0.2-30 or higher in order to upgrade to 5.4.14-28.